+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## Version
+## 0.0.2-241223d
 
+- `Update` and `Render` from upper applications are now properly called
+- Added env_logger
+- Window now has the name passed by the application
+
+## 0.0.1-240923d
 - Start of the project
 - Added github actions to build and run tests on Windows, Ubuntu and MacOS
 - Barebones entry point